Change namespace to not have `.object`. Allow `isOpVectorized` to have a `Nothing` name.
