+++ distribution/lib/Standard/Table/0.0.0-dev/src/Data/Table.enso
@@ -51,7 +51,7 @@ from project.Data.Column import get_item_string, normalize_string_for_display
 from project.Internal.Filter_Condition_Helpers import make_filter_column
 
 polyglot java import org.enso.base.ObjectComparator
-polyglot java import org.enso.table.data.column.builder.object.StorageTypeMismatch
+polyglot java import org.enso.table.data.column.builder.StorageTypeMismatch
 polyglot java import org.enso.table.data.table.Table as Java_Table
 polyglot java import org.enso.table.data.table.Column as Java_Column
 polyglot java import org.enso.table.data.table.join.Equals as Java_Join_Equals
