+++ std-bits/table/src/main/java/org/enso/table/data/column/builder/LongBuilder.java
@@ -1,9 +1,8 @@
-package org.enso.table.data.column.builder.object;
+package org.enso.table.data.column.builder;
 
 import org.enso.base.polyglot.NumericConverter;
-import org.enso.table.data.column.operation.cast.ToFloatStorageConverter;
+import org.enso.table.data.column.operation.cast.ToIntegerStorageConverter;
 import org.enso.table.data.column.storage.BoolStorage;
-import org.enso.table.data.column.storage.numeric.DoubleStorage;
 import org.enso.table.data.column.storage.numeric.LongStorage;
 import org.enso.table.data.column.storage.Storage;
 import org.enso.table.data.column.storage.type.BooleanType;
@@ -16,10 +15,10 @@ import java.util.BitSet;
 import java.util.Objects;
 
 /**
- * A builder for floating point columns.
+ * A builder for integer columns.
  */
-public class DoubleBuilder extends NumericBuilder {
-  DoubleBuilder(BitSet isMissing, long[] data, int currentSize) {
+public class LongBuilder extends NumericBuilder {
+  LongBuilder(BitSet isMissing, long[] data, int currentSize) {
     super(isMissing, data, currentSize);
   }
 
@@ -29,24 +28,28 @@ public class DoubleBuilder extends NumericBuilder {
       if (isMissing.get(i)) {
         items[i] = null;
       } else {
-        items[i] = Double.longBitsToDouble(data[i]);
+        items[i] = data[i];
       }
     }
   }
 
   @Override
   public boolean canRetypeTo(StorageType type) {
-    return false;
+    return Objects.equals(type, FloatType.FLOAT_64);
   }
 
   @Override
   public TypedBuilder retypeTo(StorageType type) {
-    throw new UnsupportedOperationException();
+    if (Objects.equals(type, FloatType.FLOAT_64)) {
+      return retypeLongBuilderToDouble(this);
+    } else {
+      throw new UnsupportedOperationException();
+    }
   }
 
   @Override
   public StorageType getType() {
-    return FloatType.FLOAT_64;
+    return IntegerType.INT_64;
   }
 
   @Override
@@ -54,41 +57,27 @@ public class DoubleBuilder extends NumericBuilder {
     if (o == null) {
       isMissing.set(currentSize++);
     } else {
-      double value = NumericConverter.coerceToDouble(o);
-      data[currentSize++] = Double.doubleToRawLongBits(value);
+      data[currentSize++] = NumericConverter.coerceToLong(o);
     }
   }
 
   @Override
   public boolean accepts(Object o) {
-    return NumericConverter.isCoercibleToDouble(o);
+    return NumericConverter.isCoercibleToLong(o);
   }
 
   @Override
   public void appendBulkStorage(Storage<?> storage) {
-    if (Objects.equals(storage.getType(), FloatType.FLOAT_64)) {
-      if (storage instanceof DoubleStorage doubleStorage) {
-        int n = doubleStorage.size();
+    if (Objects.equals(storage.getType(), IntegerType.INT_64)) {
+      if (storage instanceof LongStorage longStorage) {
+        int n = longStorage.size();
         ensureFreeSpaceFor(n);
-        System.arraycopy(doubleStorage.getRawData(), 0, data, currentSize, n);
-        BitSets.copy(doubleStorage.getIsMissing(), isMissing, currentSize, n);
+        System.arraycopy(longStorage.getRawData(), 0, data, currentSize, n);
+        BitSets.copy(longStorage.getIsMissing(), isMissing, currentSize, n);
         currentSize += n;
       } else {
         throw new IllegalStateException(
             "Unexpected storage implementation for type DOUBLE: "
-                + storage
-                + ". This is a bug in the Table library.");
-      }
-    } else if (Objects.equals(storage.getType(), IntegerType.INT_64)) {
-      if (storage instanceof LongStorage longStorage) {
-        int n = longStorage.size();
-        BitSets.copy(longStorage.getIsMissing(), isMissing, currentSize, n);
-        for (int i = 0; i < n; i++) {
-          data[currentSize++] = Double.doubleToRawLongBits((double) longStorage.getItem(i));
-        }
-      } else {
-        throw new IllegalStateException(
-            "Unexpected storage implementation for type LONG: "
                 + storage
                 + ". This is a bug in the Table library.");
       }
@@ -99,8 +88,7 @@ public class DoubleBuilder extends NumericBuilder {
           if (boolStorage.isNa(i)) {
             isMissing.set(currentSize++);
           } else {
-            double x = ToFloatStorageConverter.booleanAsDouble(boolStorage.getItem(i));
-            data[currentSize++] = Double.doubleToRawLongBits(x);
+            data[currentSize++] = ToIntegerStorageConverter.booleanAsLong(boolStorage.getItem(i));
           }
         }
       } else {
@@ -115,19 +103,26 @@ public class DoubleBuilder extends NumericBuilder {
   }
 
   /**
-   * Append a new double to this builder.
+   * Append a new integer to this builder.
    *
-   * @param data the double to append
+   * @param data the integer to append
    */
-  public void appendDouble(double data) {
+  public void appendLong(long data) {
+    int wasSize = currentSize;
+    int wasLength = this.data.length;
+
     if (currentSize >= this.data.length) {
       grow();
     }
-    appendRawNoGrow(Double.doubleToRawLongBits(data));
+
+    if (currentSize >= this.data.length) {
+      throw new IllegalStateException("currentSize=" + currentSize + "; wasSize=" + wasSize + "; wasLength=" + wasLength + "; data.length=" + this.data.length);
+    }
+    appendRawNoGrow(data);
   }
 
   @Override
-  public Storage<Double> seal() {
-    return new DoubleStorage(data, currentSize, isMissing);
+  public Storage<Long> seal() {
+    return new LongStorage(data, currentSize, isMissing);
   }
 }
