+++ std-bits/table/src/main/java/org/enso/table/data/column/builder/ObjectBuilder.java
@@ -1,4 +1,4 @@
-package org.enso.table.data.column.builder.object;
+package org.enso.table.data.column.builder;
 
 import org.enso.table.data.column.storage.ObjectStorage;
 import org.enso.table.data.column.storage.SpecializedStorage;
