+++ std-bits/table/src/main/java/org/enso/table/data/column/operation/cast/ToTextStorageConverter.java
@@ -2,7 +2,7 @@ package org.enso.table.data.column.operation.cast;
 
 import org.enso.base.Text_Utils;
 import org.enso.polyglot.common_utils.Core_Date_Utils;
-import org.enso.table.data.column.builder.object.StringBuilder;
+import org.enso.table.data.column.builder.StringBuilder;
 import org.enso.table.data.column.storage.*;
 import org.enso.table.data.column.storage.datetime.DateStorage;
 import org.enso.table.data.column.storage.datetime.DateTimeStorage;
