+++ std-bits/table/src/main/java/org/enso/table/data/column/storage/ObjectStorage.java
@@ -1,8 +1,8 @@
 package org.enso.table.data.column.storage;
 
 import java.util.BitSet;
-import org.enso.table.data.column.builder.object.Builder;
-import org.enso.table.data.column.builder.object.ObjectBuilder;
+import org.enso.table.data.column.builder.Builder;
+import org.enso.table.data.column.builder.ObjectBuilder;
 import org.enso.table.data.column.operation.map.MapOpStorage;
 import org.enso.table.data.column.operation.map.UnaryMapOperation;
 import org.enso.table.data.column.storage.type.AnyObjectType;
