+++ std-bits/table/src/main/java/org/enso/table/data/column/storage/SpecializedStorage.java
@@ -1,5 +1,6 @@
 package org.enso.table.data.column.storage;
 
+import java.util.AbstractList;
 import java.util.BitSet;
 import java.util.List;
 import org.enso.table.data.column.operation.map.MapOpStorage;
@@ -148,5 +149,28 @@ public abstract class SpecializedStorage<T> extends Storage<T> {
     }
 
     return newInstance(newData, newSize);
+  }
+
+  @Override
+  public List<Object> toList() {
+    return new ReadOnlyList<>(this);
+  }
+
+  private class ReadOnlyList<S> extends AbstractList<Object> {
+    private final SpecializedStorage<S> storage;
+
+    public ReadOnlyList(SpecializedStorage<S> storage) {
+      this.storage = storage;
+    }
+
+    @Override
+    public Object get(int index) {
+      return storage.getItemBoxed(index);
+    }
+
+    @Override
+    public int size() {
+      return storage.size();
+    }
   }
 }
