+++ std-bits/table/src/main/java/org/enso/table/data/column/storage/datetime/DateStorage.java
@@ -1,8 +1,8 @@
 package org.enso.table.data.column.storage.datetime;
 
 import java.time.LocalDate;
-import org.enso.table.data.column.builder.object.Builder;
-import org.enso.table.data.column.builder.object.DateBuilder;
+import org.enso.table.data.column.builder.Builder;
+import org.enso.table.data.column.builder.DateBuilder;
 import org.enso.table.data.column.operation.map.MapOpStorage;
 import org.enso.table.data.column.operation.map.UnaryIntegerOp;
 import org.enso.table.data.column.operation.map.datetime.DateTimeIsInOp;
