+++ std-bits/table/src/main/java/org/enso/table/data/column/storage/numeric/AbstractLongStorage.java
@@ -1,8 +1,8 @@
 package org.enso.table.data.column.storage.numeric;
 
 import java.util.BitSet;
-import org.enso.table.data.column.builder.object.Builder;
-import org.enso.table.data.column.builder.object.NumericBuilder;
+ import org.enso.table.data.column.builder.Builder;
+import org.enso.table.data.column.builder.NumericBuilder;
 import org.enso.table.data.column.operation.map.MapOpStorage;
 import org.enso.table.data.column.operation.map.MapOperationProblemBuilder;
 import org.enso.table.data.column.operation.map.UnaryLongToLongOp;
