+++ std-bits/table/src/main/java/org/enso/table/data/table/Column.java
@@ -2,7 +2,7 @@ package org.enso.table.data.table;
 
 import org.enso.base.Text_Utils;
 import org.enso.base.polyglot.Polyglot_Utils;
-import org.enso.table.data.column.builder.object.InferredBuilder;
+import org.enso.table.data.column.builder.InferredBuilder;
 import org.enso.table.data.column.storage.BoolStorage;
 import org.enso.table.data.column.storage.Storage;
 import org.enso.table.data.index.DefaultIndex;
