+++ std-bits/table/src/main/java/org/enso/table/parsing/DateParser.java
@@ -2,8 +2,8 @@ package org.enso.table.parsing;
 
 import java.util.Locale;
 import org.enso.polyglot.common_utils.Core_Date_Utils;
-import org.enso.table.data.column.builder.object.Builder;
-import org.enso.table.data.column.builder.object.DateBuilder;
+import org.enso.table.data.column.builder.Builder;
+import org.enso.table.data.column.builder.DateBuilder;
 
 public class DateParser extends BaseTimeParser {
   public DateParser(String[] formats, Locale locale) {
