+++ std-bits/table/src/main/java/org/enso/table/parsing/DateTimeParser.java
@@ -2,8 +2,8 @@ package org.enso.table.parsing;
 
 import java.util.Locale;
 import org.enso.polyglot.common_utils.Core_Date_Utils;
-import org.enso.table.data.column.builder.object.Builder;
-import org.enso.table.data.column.builder.object.DateTimeBuilder;
+import org.enso.table.data.column.builder.Builder;
+import org.enso.table.data.column.builder.DateTimeBuilder;
 
 public class DateTimeParser extends BaseTimeParser {
   public DateTimeParser(String[] formats, Locale locale) {
