+++ std-bits/table/src/main/java/org/enso/table/parsing/IdentityParser.java
@@ -1,7 +1,7 @@
 package org.enso.table.parsing;
 
 import java.util.List;
-import org.enso.table.data.column.builder.object.StringBuilder;
+import org.enso.table.data.column.builder.StringBuilder;
 import org.enso.table.data.column.storage.Storage;
 import org.enso.table.parsing.problems.ProblemAggregator;
 import org.enso.table.problems.WithProblems;
