+++ std-bits/table/src/main/java/org/enso/table/parsing/TimeOfDayParser.java
@@ -2,8 +2,8 @@ package org.enso.table.parsing;
 
 import java.time.LocalTime;
 import java.util.Locale;
-import org.enso.table.data.column.builder.object.Builder;
-import org.enso.table.data.column.builder.object.TimeOfDayBuilder;
+import org.enso.table.data.column.builder.Builder;
+import org.enso.table.data.column.builder.TimeOfDayBuilder;
 
 public class TimeOfDayParser extends BaseTimeParser {
   public TimeOfDayParser(String[] formats, Locale locale) {
