+++ std-bits/table/src/main/java/org/enso/table/parsing/WhitespaceStrippingParser.java
@@ -1,6 +1,6 @@
 package org.enso.table.parsing;
 
-import org.enso.table.data.column.builder.object.Builder;
+import org.enso.table.data.column.builder.Builder;
 import org.enso.table.parsing.problems.ProblemAggregator;
 
 /**
